# public/swagger.yaml
components:
...
    schemas:
        SignupResponse:
            properties:
                name:
                    type: string
                email:
                    type: string
            required:
                - name
                - email
            type: object
            additionalProperties: false
        Role:
            enum:
                - ADMIN
                - SHELTER
                - USER
            type: string
        UserPayload:
            properties:
                name:
                    type: string
                email:
                    type: string
                password:
                    type: string
                role:
                    $ref: '#/components/schemas/Role'
            required:
                - name
                - email
                - password
                - role
            type: object
            additionalProperties: false
            example:
                name: 'Jack Doe'
                email: johndoe@example.com
                password: '123456'
                role: USER
        EmailChangeRequest:
            enum:
                - currentEmailStep
                - newEmailStep
            type: string
        SendCodePayload:
            properties:
                email:
                    type: string
                    description: 'Email of user'
                    example: johndoe@example.com
                emailChangeRequest:
                    $ref: '#/components/schemas/EmailChangeRequest'
            required:
                - email
            type: object
            additionalProperties: false
        VerificationResponse:
            properties:
                tokens:
                    properties:
                        refreshToken:
                            type: string
                        accessToken:
                            type: string
                    required:
                        - refreshToken
                        - accessToken
                    type: object
                isVerified:
                    type: boolean
            required:
                - tokens
                - isVerified
            type: object
            additionalProperties: false
        VerificationPayload:
            properties:
                email:
                    type: string
                    description: 'Email of user'
                    example: johndoe@example.com
                verificationCode:
                    type: string
                    description: 'Verification code of user'
                    example: '654321'
            required:
                - email
                - verificationCode
            type: object
            additionalProperties: false
        TokenResponse:
            properties:
                tokens:
                    properties:
                        refreshToken:
                            type: string
                        accessToken:
                            type: string
                    required:
                        - refreshToken
                        - accessToken
                    type: object
            required:
                - tokens
            type: object
            additionalProperties: false
        LoginPayload:
            properties:
                email:
                    type: string
                    description: 'Email of user'
                    example: johndoe@example.com
                password:
                    type: string
                    description: 'Password for user'
                    example: '123456'
            required:
                - email
                - password
            type: object
            additionalProperties: false
        EmailPayload:
            properties:
                email:
                    type: string
                    description: 'Email of user'
                    example: johndoe@example.com
            required:
                - email
            type: object
            additionalProperties: false
        CheckPasswordPayload:
            properties:
                password:
                    type: string
                    description: 'Password for user'
                    example: '123456'
            required:
                - password
            type: object
            additionalProperties: false
        ResetPasswordPayload:
            properties:
                email:
                    type: string
                    description: 'Email of user'
                    example: johndoe@example.com
                newPassword:
                    type: string
                    description: 'Password for user'
                    example: '123456'
            required:
                - email
                - newPassword
            type: object
            additionalProperties: false
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
info:
    title: express-ts
    version: 1.0.0
    license:
        name: ISC
    contact: {}
openapi: 3.0.0
paths:
    /auth/signup:
        post:
            operationId: Signup
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SignupResponse'
                            examples:
                                'Example 1':
                                    value: {name: 'John Doe', email: johndoe@example.com}
            summary: 'Accepts user info, creates user and returns user info except password'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserPayload'
    /auth/verificationCode:
        post:
            operationId: SendVerificationCode
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Generates a 6-digit code and sends it'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SendCodePayload'
    /auth/verifyEmail:
        post:
            operationId: VerifyEmail
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/VerificationResponse'
                            examples:
                                'Example 1':
                                    value: {isVerified: true, tokens: {accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..., refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...}}
            summary: 'Accepts a 6-digit code from user, verifies the code in db'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/VerificationPayload'
    /auth/login:
        post:
            operationId: Login
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TokenResponse'
                            examples:
                                'Example 1':
                                    value: {tokens: {accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..., refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...}}
            summary: 'Verifies the user''s email and password and returns JWT tokens'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginPayload'
    /auth/refresh:
        post:
            operationId: Refresh
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TokenResponse'
                            examples:
                                'Example 1':
                                    value: {tokens: {accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..., refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...}}
            summary: 'Refreshes access token'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
    /auth/updateProfile:
        put:
            operationId: UpdateProfile
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
                            examples:
                                'Example 1':
                                    value: {name: 'John Doe', address: '123 Main St', bio: 'Pet lover with years of experience fostering cats and dogs. I have a spacious home with a large, secure backyard. I love active pets and have a preference for medium to large dogs.', profilePhoto: 'https://example.com/userProfilePhoto.jpg', removeProfilePhoto: true}
            summary: 'Updates user''s profile'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: false
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                address:
                                    type: string
                                bio:
                                    type: string
                                profilePhoto:
                                    type: string
                                    format: binary
    /auth/checkEmail:
        get:
            operationId: CheckEmail
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Checks if user''s new email already has a linked account'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    description: 'New email for user'
                    in: query
                    name: email
                    required: true
                    schema:
                        type: string
    /auth/changeEmail:
        put:
            operationId: ChangeEmail
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TokenResponse'
                            examples:
                                'Example 1':
                                    value: {tokens: {accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..., refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...}}
            summary: 'Changes user''s email'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EmailPayload'
    /auth/checkPassword:
        post:
            operationId: CheckPassword
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Checks if user''s entered password is correct'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CheckPasswordPayload'
    /auth/changePassword:
        put:
            operationId: ChangePassword
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Changes user''s password'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CheckPasswordPayload'
    /auth/logout:
        delete:
            operationId: Logout
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Removes JWT tokens and returns success message'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
    /auth/requestPasswordReset:
        post:
            operationId: RequestPasswordReset
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Checks if user''s entered email has an associated account and create reset tokens'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EmailPayload'
    /auth/verifyResetToken:
        get:
            operationId: VerifyResetToken
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    email: {}
                                required:
                                    - email
                                type: object
            summary: 'Verifies the reset password token for user'
            tags:
                - Auth
            security: []
            parameters:
                -
                    in: query
                    name: resetToken
                    required: true
                    schema:
                        type: string
    /auth/resetPassword:
        put:
            operationId: ResetPassword
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Changes user''s password'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ResetPasswordPayload'
    /auth/getAlternateEmail:
        post:
            operationId: GetAlternateEmail
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    message: {type: string}
                                    code: {type: number, format: double}
                                required:
                                    - message
                                    - code
                                type: object
            summary: 'Sends email to user to provide alternate email to sign up as shelter'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EmailPayload'
...
servers:
    -
        url: /
...